requirements.txt
# CPU‑only сборка
transformers
accelerate
pillow
bitsandbytes
faster-whisper
ctranslate2
gradio
torch
einops
timm
sentencepiece